requirements.txt
aiogram>=3.1.1,<4.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dotenv>=0.19.0
aiohttp>=3.8.0 